updated and changed tests for call insertion
